+++ CHANGELOG.md
@@ -5,7 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) once
 we hit release version 1.0.0.
 
-## [0.15.2] - YYYY-MM-DD
+
+## [0.15.3] - YYYY-MM-DD
+
+
+## [0.15.2] - 2024-11-06
 
 ### Added
 - added `timesSileSiesta` which can read siesta TIMES output